# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TooDue API
  version: 1.0.0
  summary: TooDue task, project, collaboration, and AI-assistant API.
  description: |
    TooDue has one canonical API surface under `/api`.

    The web app and external integrations call the same resource endpoints. Browser sessions authenticate with the `toodue_session` cookie; external integrations should authenticate with `Authorization: Bearer tdue_...` API keys created in the TooDue app.

    The legacy `/api/ai/*` aliases remain for compatibility, but new clients should use the canonical endpoints documented here.
  contact:
    name: TooDue
    url: https://toodue.com
  license:
    name: Proprietary
    url: https://toodue.com
servers:
- url: https://app.toodue.com/api
  description: Production API
...
security:
- BearerApiKey: []
- SessionCookie: []
paths:
...
  "/projects":
    get:
      tags:
      - Public API
      summary: List projects visible to the current user
      operationId: listProjects
      responses:
        '200':
          description: Projects including member lists.
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ProjectWithMembers"
        '401':
          "$ref": "#/components/responses/Unauthorized"
    post:
      tags:
      - Public API
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateProject"
            examples:
              default:
                value:
                  name: Follow up
                  color: blue
                  parent_id: 
      responses:
        '200':
          description: Created project.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ProjectWithMembers"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
...
  "/tasks":
    get:
      tags:
      - Public API
      summary: List active or completed tasks
      operationId: listTasks
      parameters:
      - "$ref": "#/components/parameters/ProjectIdQuery"
      - name: completed
        in: query
        description: When true with `project_id`, returns completed tasks for that project. Otherwise returns active tasks.
        schema:
          type: boolean
      responses:
        '200':
          description: Matching tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Task"
        '401':
          "$ref": "#/components/responses/Unauthorized"
    post:
      tags:
      - Public API
      summary: Create a task or subtask
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTask"
            examples:
              task:
                value:
                  project_id: 1
                  name: Call the plumber
                  description: Ask about the upstairs sink.
                  due_date: '2026-07-07'
                  due_time: '14:30'
                  deadline: '2026-07-10'
                  priority: 3
      responses:
        '200':
          description: Created task.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Task"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
...
  "/tasks/{id}":
    get:
      tags:
      - Public API
      summary: Get task details
      operationId: getTask
      parameters:
      - "$ref": "#/components/parameters/TaskId"
      responses:
        '200':
          description: Task plus subtasks, comments, and attachments.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TaskDetail"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
    patch:
      tags:
      - Public API
      summary: Update a task
      operationId: updateTask
      parameters:
      - "$ref": "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateTask"
            examples:
              complete:
                value:
                  completed: true
              reschedule:
                value:
                  due_date: '2026-07-08'
                  due_time: 
      responses:
        '200':
          description: Updated task.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Task"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
    delete:
      tags:
      - Public API
      summary: Delete a task
      operationId: deleteTask
      parameters:
      - "$ref": "#/components/parameters/TaskId"
      responses:
        '200':
          "$ref": "#/components/responses/Ok"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: tdue API key
      description: API keys start with `tdue_` and are created in the TooDue app.
    SessionCookie:
      type: apiKey
      in: cookie
      name: toodue_session
      description: Browser session cookie used by the TooDue web app.
  parameters:
...
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ProjectIdQuery:
      name: project_id
      in: query
      schema:
        type: integer
        format: int64
      description: Restrict results to one project.
  responses:
    Ok:
      description: Operation succeeded.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Ok"
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user cannot access the requested resource.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
  schemas:
    Attachment:
      type: object
      required:
      - id
      - task_id
      - user_id
      - filename
      - mime
      - size
      - created_at
      properties:
        id:
          type: integer
          format: int64
        task_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        filename:
          type: string
        mime:
          type: string
        size:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    Comment:
      type: object
      required:
      - id
      - task_id
      - user_id
      - user_name
      - body
      - created_at
      properties:
        id:
          type: integer
          format: int64
        task_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        user_name:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
    CreateProject:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        color:
          type: string
          default: slate
        parent_id:
          type:
          - integer
          - 'null'
          format: int64
    CreateTask:
      type: object
      required:
      - name
      properties:
        project_id:
          type: integer
          format: int64
          description: Required unless `parent_id` is provided.
        parent_id:
          type: integer
          format: int64
          description: Parent task id when creating a subtask. Inherits the parent project.
        name:
          type: string
        description:
          type: string
          default: ''
        due_date:
          type:
          - string
          - 'null'
          format: date
        due_time:
          type:
          - string
          - 'null'
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
        deadline:
          type:
          - string
          - 'null'
          format: date
        priority:
          type: integer
          minimum: 1
          maximum: 4
          default: 4
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          examples:
          - unauthorized
    Member:
      type: object
      required:
      - id
      - name
      - email
      - role
      - project_id
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
          - owner
          - member
        project_id:
          type: integer
          format: int64
    Ok:
      type: object
      required:
      - ok
      properties:
        ok:
          type: boolean
          const: true
    Project:
      type: object
      required:
      - id
      - name
      - color
      - owner_id
      - is_inbox
      - sort_order
      - created_at
      - active_count
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        color:
          type: string
          examples:
          - slate
          - blue
          - emerald
          - rose
        parent_id:
          type:
          - integer
          - 'null'
          format: int64
        owner_id:
          type: integer
          format: int64
        is_inbox:
          type: integer
          description: 1 for the built-in Inbox, otherwise 0.
          enum:
          - 0
          - 1
        sort_order:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        active_count:
          type: integer
          format: int64
    ProjectWithMembers:
      allOf:
      - "$ref": "#/components/schemas/Project"
      - type: object
        required:
        - members
        properties:
          members:
            type: array
            items:
              "$ref": "#/components/schemas/Member"
    Task:
      type: object
      required:
      - id
      - project_id
      - creator_id
      - name
      - description
      - priority
      - sort_order
      - created_at
      - updated_at
      - comment_count
      - attachment_count
      - subtask_count
      - subtask_done_count
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        parent_id:
          type:
          - integer
          - 'null'
          format: int64
        creator_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        due_date:
          type:
          - string
          - 'null'
          format: date
        due_time:
          type:
          - string
          - 'null'
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          examples:
          - '14:30'
        deadline:
          type:
          - string
          - 'null'
          format: date
        priority:
          type: integer
          minimum: 1
          maximum: 4
          description: 1 is highest priority; 4 is normal/default.
        completed_at:
          type:
          - string
          - 'null'
          format: date-time
        sort_order:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        comment_count:
          type: integer
          format: int64
        attachment_count:
          type: integer
          format: int64
        subtask_count:
          type: integer
          format: int64
        subtask_done_count:
          type: integer
          format: int64
    TaskDetail:
      type: object
      required:
      - task
      - subtasks
      - comments
      - attachments
      properties:
        task:
          "$ref": "#/components/schemas/Task"
        subtasks:
          type: array
          items:
            "$ref": "#/components/schemas/Task"
        comments:
          type: array
          items:
            "$ref": "#/components/schemas/Comment"
        attachments:
          type: array
          items:
            "$ref": "#/components/schemas/Attachment"
    UpdateTask:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        due_date:
          type:
          - string
          - 'null'
          format: date
          description: Send null to clear.
        due_time:
          type:
          - string
          - 'null'
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          description: Send null to clear.
        deadline:
          type:
          - string
          - 'null'
          format: date
          description: Send null to clear.
        priority:
          type: integer
          minimum: 1
          maximum: 4
        completed:
          type: boolean
          description: true completes the task; false reopens it.
        project_id:
          type: integer
          format: int64
          description: Move the task to another project.
...